==> Rcode/01_source_functions.R <==
# 01_source_functions.R
# get all functions into memory

code.dir <- ".\\Rcode\\"
source(paste0(code.dir,"Read.ASAP3.dat.file.R"))
source(paste0(code.dir,"adjust.asap.R"))
source(paste0(code.dir,"Write.ASAP3.dat.file.R"))
source(paste0(code.dir,"run.retro.mults.R"))
source(paste0(code.dir,"calc.SSBrho.R"))
source(paste0(code.dir,"clean.up.files.R"))
